Move widget traits to new module
